refactor, feat: refactors main code from PageBrowser into searchInput.tsx file for better scoping and variable changing; implemented search-functionality such that when a new search is generated, the currently existing cards are deleted and then updated to the new cards from the input query
