+++ src/components/browseMovies/SearchInput.tsx
@@ -1,5 +1,6 @@
 import { useState } from "react"
 import Card from "../cards/Card"
+import BrowsingSubheader from "./BrowsingSubheader";
 
 const SearchInput = () => {
 
@@ -11,6 +12,43 @@ const SearchInput = () => {
     // causing the browsing-sub-sub title to change to whatever the input string was (reformatted)
     // and the cards to display like how they are displayed in the home trending page.  
 
+    let [results, setResults] = useState([]);
+
+    
+    const urlTrending = 'https://api.themoviedb.org/3/trending/movie/week?language=en-US';
+    const options = {
+        method: 'GET',
+        headers: {
+            accept: 'application/json',
+            Authorization: 'Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiI1YWIxNjYwNWY5MzQ5MzJiYzk4N2NlZTJjYjRiZGMwYyIsInN1YiI6IjY2MmNmMTk1ZTMzZjgzMDEyMjIxMDc3NyIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.Xu5UsO6VS90CsemI07dMUU6xjKcbTxQ1UrfGjC9iFjc'
+        }
+    }; 
+
+    // if the call was succcesful then store the data gathered inside the results array
+    if (results.length == 0) {
+        fetch(urlTrending, options)
+            .then((res: any) => res.json())
+            .then((data: any) => {
+                console.log(data);
+                setResults(data.results);
+            })
+            .catch((err: any) => console.error('error:' + err));
+
+    }
+
+    // movieCards element maps the collected elements from the array collected
+    // by the API and is used in return code through <Card> element
+
+    let movieCards = results.map((movie: any) =>
+    (
+        <li key={movie.id} className="trending-list">
+            <Card movie={movie} />
+        </li>
+
+    )
+    )
+
+
     let [inputValue, setInputValue] = useState("")
 
     const inputChange = (e: any) => {
@@ -18,9 +56,6 @@ const SearchInput = () => {
     }
 
 
-    let [results, setResults] = useState([]);
-
-    
     const searchClick = (e: any) => {
 
         if (inputValue != "") {
@@ -29,6 +64,7 @@ const SearchInput = () => {
             // generate query based on input value
 
             //TODO
+            getSpecificQuery();
 
             // changes input box back to empty string
 
@@ -42,6 +78,42 @@ const SearchInput = () => {
     // 'main content'ish and then port the logic for initial trendiing query here; once the user inputs
     // a new query, the cards update based on the user's input query
 
+    const getSpecificQuery = () =>{
+
+        const urlSpecific = `https://api.themoviedb.org/3/search/movie?query=${inputValue}&include_adult=false&language=en&page=1`;
+        const options = {
+            method: 'GET',
+            headers: {
+                accept: 'application/json',
+                Authorization: 'Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiI1YWIxNjYwNWY5MzQ5MzJiYzk4N2NlZTJjYjRiZGMwYyIsInN1YiI6IjY2MmNmMTk1ZTMzZjgzMDEyMjIxMDc3NyIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.Xu5UsO6VS90CsemI07dMUU6xjKcbTxQ1UrfGjC9iFjc'
+            }
+        }; 
+    
+        // if the call was succcesful then store the data gathered inside the results array
+            fetch(urlSpecific, options)
+                .then((res: any) => res.json())
+                .then((data: any) => {
+                    console.log(data);
+                    setResults([])
+                    setResults(data.results);
+                })
+                .catch((err: any) => console.error('error:' + err));
+
+            // movieCards element maps the collected elements from the array collected
+    // by the API and is used in return code through <Card> element
+
+    movieCards = results.map((movie: any) =>
+        (
+            <li key={movie.id} className="trending-list">
+                <Card movie={movie} />
+            </li>
+    
+        )
+        )
+    
+    
+    }
+
     const setTextChange = () => {
 
         const text = document.getElementById("subheader-title-inner-content")
@@ -52,11 +124,26 @@ const SearchInput = () => {
 
 
     return (
-        <div className="search-input-wrapper">
-            <input id="search-bar" value={inputValue} className="search-input-editor" type="text" placeholder="Browse for Movies..." onChange={inputChange} />
-            <button className="search-input-button" onClick={searchClick} >
-                <i className="icon fa-solid fa-magnifying-glass"></i>
-            </button>
+        <div className=" total-container">
+            <div className="browsing-container">
+
+                <div className="search-input-wrapper">
+                    <input id="search-bar" value={inputValue} className="search-input-editor" type="text" placeholder="Browse for Movies..." onChange={inputChange} />
+                    <button className="search-input-button" onClick={searchClick} >
+                        <i className="icon fa-solid fa-magnifying-glass"></i>
+                    </button>
+                </div>
+            </div>
+
+            <hr className="browsing-break" />
+            <div className="browsing-card-container">
+                <BrowsingSubheader browsingSubheaderTitle="Current Trending Movies" />
+                <div className="trending-main-content">
+                    {movieCards}
+                </div>
+            </div>
+
+
         </div>
     )
 }
